tests/spellchecker_requirements.txt
rich
pyspellchecker
